added score function and (eventually) got newLevel working
to check when a new level is required, I'd set it to check the status of every block in a for loop - took ages, so changed it so that for every hit it added one to blocksthathavebeenhit so there's just one if statement to check. lots of global variable problems with the rest of newLevel!